requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30
weasyprint==60.1
flask-share==0.1.1
flask-caching==2.1.0 